Added iterative peptide analysis
I am seeing a significant change in the outputs. Works by filtering the gmt with only the most the significant peptide found in each iteration that has not been found in the previous iterations. Current saves the final gmt to a file named "final.gmt", but this could be a parameter.